[CAL][10] - Implement Nonces and GetChainFeePriceUpdate
